fix(notifications): fixed bugs with signals
